# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ToolsHub API - Tool Rental Management Platform",
    "version": "1.0.0",
    "description": "API documentation for the ToolsHub application with CRUD operations for tools and rentals."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Tool": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "Power_tools",
              "Hand_tools",
              "Gardening",
              "Painting",
              "Other"
            ]
          },
          "hourlyRate": {
            "type": "number"
          },
          "depositAmount": {
            "type": "number"
          },
          "quantity": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "Available",
              "Rented",
              "Maintenance"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateToolDTO": {
        "type": "object",
        "required": [
          "name",
          "description",
          "category",
          "hourlyRate",
          "depositAmount",
          "quantity"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "category": {
            "type": "string",
            "enum": [
              "Power_tools",
              "Hand_tools",
              "Gardening",
              "Painting",
              "Other"
            ]
          },
          "hourlyRate": {
            "type": "number",
            "minimum": 0.01
          },
          "depositAmount": {
            "type": "number",
            "minimum": 1
          },
          "quantity": {
            "type": "integer",
            "minimum": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "Available",
              "Rented",
              "Maintenance"
            ]
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "code": {
                "type": "string"
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/tools": {
      "get": {
        "summary": "Get all tools",
        "description": "Retrieve a list of all available tools",
        "tags": [
          "Tools"
        ],
        "responses": {
          "200": {
            "description": "Tools retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Tool"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new tool",
        "description": "Add a new tool to the inventory",
        "tags": [
          "Tools"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateToolDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tool created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Tool"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/tools/category/{category}": {
      "get": {
        "summary": "Get tools by category",
        "description": "Retrieve all tools in a specific category",
        "tags": [
          "Tools"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "power_tools",
                "hand_tools",
                "gardening",
                "painting",
                "other"
              ]
            },
            "description": "Tool category to filter by"
          }
        ],
        "responses": {
          "200": {
            "description": "Tools retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Tool"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tools/{id}": {
      "get": {
        "summary": "Get tool by ID",
        "description": "Retrieve a single tool by its unique identifier",
        "tags": [
          "Tools"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Tool ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Tool retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Tool"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tool not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a tool",
        "description": "Update an existing tool's information",
        "tags": [
          "Tools"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Tool ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "power_tools",
                      "hand_tools",
                      "gardening",
                      "painting",
                      "other"
                    ]
                  },
                  "hourlyRate": {
                    "type": "number"
                  },
                  "depositAmount": {
                    "type": "number"
                  },
                  "quantity": {
                    "type": "integer"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "Available",
                      "Rented",
                      "Maintenance"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tool updated successfully"
          },
          "404": {
            "description": "Tool not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a tool",
        "description": "Remove a tool from the inventory",
        "tags": [
          "Tools"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Tool ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Tool deleted successfully"
          },
          "404": {
            "description": "Tool not found"
          }
        }
      }
    }
  },
...
}